+++ python_search/init/install_dependencies.py
@@ -1,6 +1,6 @@
 import os
 
-from python_search.environment import is_mac
+from python_search.environment import is_mac, is_archlinux
 
 
 class InstallDependencies:
@@ -16,9 +16,11 @@ class InstallDependencies:
         self._install_fzf()
         self._install_kitty()
         self._install_ack()
-        self._install_tk_mac()
+        self._install_tk()
         self._install_zsh_mac()
         self._install_shortcut_mac()
+        self._install_wctrl()
+        self._install_xsel()
 
         print("""
 Installation successful!
@@ -34,12 +36,13 @@ Installation successful!
                 "Dont know how to install ack for your platform, please do so manually"
             )
 
-    def _install_tk_mac(self):
+    def _install_tk(self):
 
-        if not is_mac():
-            return
+        if is_mac():
+            os.system("brew install python-tk")
+        if is_archlinux():
+            os.system("sudo pacman -S tk")
 
-        os.system("brew install python-tk")
 
     def _install_zsh_mac(self):
 
@@ -109,3 +112,10 @@ Installation successful!
 
         # sets a light theme in kitty
         os.system("kitty +kitten themes --reload-in=all One Half Light")
+
+    def _install_wctrl(self):
+        if is_archlinux():
+            os.system("sudo pacman -S wmctrl")
+    def _install_xsel(self):
+        if is_archlinux():
+            os.system("sudo pacman -S xsel")